able to loop over multiple visits
